all weapon upgrades interchangeable

+++ game/js/main.js
@@ -71,6 +71,7 @@ function preload() {
     game.load.spritesheet('orbs', 'img/Effect95.png', 128, 128)
     game.load.spritesheet('gun2', 'img/sparkling-fireball-wind.png', 256, 256)
     game.load.spritesheet('gun3', 'img/lasers.png', 512, 128)
+    game.load.spritesheet('gun4', 'img/sparkling-fireball-small-wind.png', 256, 256)
     // Load effects images
     game.load.spritesheet('aura', 'img/Aura38.png', 128, 128)
     game.load.spritesheet('explosion', 'img/Explosion03.png', 128, 128)
@@ -105,6 +106,7 @@ function create() {
     bluePlasma = addGroup(200, 'bluePlasma')
     gun2s = addGroup(200, 'gun2')
     gun3s = addGroup(30, 'gun3')
+    gun4s = addGroup(600, 'gun4')
     enemyBluePlasma = addGroup(400, "bluePlasma")
     orbs = addGroup(200, "orbs")
     // Create effects
@@ -120,9 +122,9 @@ function create() {
     
     // Create powerups
     upgrade = addGroup(5, "upgrade")
-    ship2s = addGroup(2, "ship2")
-    ship3s = addGroup(2, "ship3")
-    ship4s = addGroup(2, "ship4")
+    ship2s = addGroup(200, "ship2")
+    ship3s = addGroup(200, "ship3")
+    ship4s = addGroup(200, "ship4")
 
     cursors = game.input.keyboard.createCursorKeys();
     game.input.keyboard.addKeyCapture([Phaser.Keyboard.SPACEBAR, Phaser.Keyboard.ENTER]);
@@ -155,11 +157,11 @@ function update() {
         }
     })
     // Check for collision deaths
-    // game.physics.arcade.overlap(ship, alienSpaceShips, gameOver, null, this);
-    // game.physics.arcade.overlap(ship, alienMediumShips, gameOver, null, this);
-    // game.physics.arcade.overlap(ship, alienCannons, gameOver, null, this);
-    // game.physics.arcade.overlap(ship, enemyBluePlasma, gameOver, null, this);
-    // game.physics.arcade.overlap(ship, orbs, gameOver, null, this);
+    game.physics.arcade.overlap(ship, alienSpaceShips, gameOver, null, this);
+    game.physics.arcade.overlap(ship, alienMediumShips, gameOver, null, this);
+    game.physics.arcade.overlap(ship, alienCannons, gameOver, null, this);
+    game.physics.arcade.overlap(ship, enemyBluePlasma, gameOver, null, this);
+    game.physics.arcade.overlap(ship, orbs, gameOver, null, this);
     // Check for upgraded ship parts collisions
     game.physics.arcade.overlap(ship2s, alienSpaceShips, downgradeShip, null, this);
     game.physics.arcade.overlap(ship2s, alienMediumShips, downgradeShip, null, this);
@@ -192,6 +194,10 @@ function update() {
     game.physics.arcade.overlap(gun3s, alienSpaceShips, damageSmallShips, null, this);
     game.physics.arcade.overlap(gun3s, alienMediumShips, damageMediumShips, null, this);
     game.physics.arcade.overlap(gun3s, alienCannons, damageCannons, null, this);
+
+    game.physics.arcade.overlap(gun4s, alienSpaceShips, damageSmallShips, null, this);
+    game.physics.arcade.overlap(gun4s, alienMediumShips, damageMediumShips, null, this);
+    game.physics.arcade.overlap(gun4s, alienCannons, damageCannons, null, this);
 
     // Check for upgrade collisions
     game.physics.arcade.overlap(ship, upgrade, collectPowerUp, null, this);
@@ -339,13 +345,14 @@ function shootWeapon() {
             [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 
             11, 12, 13, 14, 15, 16, 17, 18, 19, 
             20, 21, ,22, 23, 24, 25, 26, 27, 28, 29, 
-            30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42], 
+            30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 
+            41, 42, 43, 44, 45, 46,], 
             20, true);
         gun2.animations.play('shoot')
         gun2.anchor.setTo(.5)
         gun2.body.setSize(5, 5, 0, 0)
         gun2.body.velocity.x = 1000
-        gun2sReset = game.time.now + 100
+        gun2sReset = game.time.now + 200
     } else if (currentWeapon === "gun3") {
         var gun3 = gun3s.getFirstExists(false)
         if (game.time.now < gun3sReset) {
@@ -361,7 +368,27 @@ function shootWeapon() {
         gun3.anchor.setTo(.5)
         gun3.body.setSize(5, 5, 0, 0)
         gun3.body.velocity.x = 1500
-        gun3sReset = game.time.now + 500
+        gun3sReset = game.time.now + 750
+    } else if (currentWeapon === "gun4") {
+        var gun4 = gun4s.getFirstExists(false)
+        if (game.time.now < gun4sReset) {
+            return;
+        }
+        gun4.reset(ship.x + 60, ship.y)
+        gun4.angle = -90
+        gun4.scale.setTo(.3, .3)
+        gun4.animations.add('shoot', 
+            [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 
+            11, 12, 13, 14, 15, 16, 17, 18, 19, 
+            20, 21, ,22, 23, 24, 25, 26, 27, 28, 29, 
+            30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40,
+            41, 42, 43, 44, 45, 46], 
+            20, true);
+        gun4.animations.play('shoot')
+        gun4.anchor.setTo(.5)
+        gun4.body.setSize(5, 5, 0, 0)
+        gun4.body.velocity.x = 600
+        gun4sReset = game.time.now + 10
     }
 } 
 
@@ -371,15 +398,17 @@ function damageSmallShips(projectile, alienSpaceShip) {
     dmg.scale.setTo(.75, .75)
     dmg.animations.play('dmg');
     dmg.anchor.setTo(.5);
-    projectile.kill();
+    if (currentWeapon != "gun4") {
+        projectile.kill();
+    }
     if (currentWeapon === "plasma") {
         alienSpaceShip.life -= 100
     } else if (currentWeapon === "gun2") {
-        alienSpaceShip.life -= 80
+        alienSpaceShip.life -= 200
     } else if (currentWeapon === "gun3") {
         alienSpaceShip.life -= 600
     } else if (currentWeapon === "gun4") {
-        alienSpaceShip.life -= 200
+        alienSpaceShip.life -= 15
     }
     if (alienSpaceShip.life <= 0) {
         score += 100
@@ -397,11 +426,11 @@ function damageMediumShips(projectile, alienMediumShip) {
     if (currentWeapon === "plasma") {
         alienMediumShip.life -= 100
     } else if (currentWeapon === "gun2"){
-        alienMediumShip.life -= 80
+        alienMediumShip.life -= 200
     } else if (currentWeapon === "gun3") {
         alienMediumShip.life -= 600
     } else if (currentWeapon === "gun4") {
-        alienMediumShip.life -= 200
+        alienMediumShip.life -= 15
     }
     if (alienMediumShip.life <= 0) {
         dmg.reset(alienMediumShip.x, alienMediumShip.y);
@@ -423,17 +452,17 @@ function damageCannons(projectile, alienCannon) {
     if (currentWeapon === "plasma") {
         alienCannon.life -= 100
     } else if (currentWeapon === "gun2"){
-        alienCannon.life -= 80
+        alienCannon.life -= 200
     } else if (currentWeapon === "gun3") {
         alienCannon.life -= 600
     } else if (currentWeapon === "gun4") {
-        alienCannon.life -= 400
+        alienCannon.life -= 15
     }
     if (alienCannon.life <= 0) {
         dmg.reset(alienCannon.x, alienCannon.y)
         dmg.scale.set(2.2, 2.2)
         dmg.animations.play('dmg')
-        score += 10000
+        score += 5000
         alienCannon.kill();
     }
 }
@@ -451,8 +480,7 @@ function gameOver(ship, enemy) {
 } 
 
 function dropPowerUp(alienMediumShip) {
-    var chance = 1
-    // Math.ceil(Math.random() * 3)
+    var chance = Math.ceil(Math.random() * 3)
     if (chance === 1) {
         var pUp = upgrade.getFirstExists(false);
         pUp.reset(alienMediumShip.x, alienMediumShip.y);
@@ -468,13 +496,12 @@ function collectPowerUp(ship, upgrade) {
     console.log('num children', ship.children.length)
     if(shipUpgrade){
         shipUpgrade.kill()
-        // ship2s.addChild(shipUpgrade)
-        ship3s.addChild(shipUpgrade)
-        // ship4s.addChild(shipUpgrade)
+    //     // ship2s.addChild(shipUpgrade)
+    //     ship3s.addChild(shipUpgrade)
+    //     // ship4s.addChild(shipUpgrade)
     }
     
-    var chance = 2
-    // Math.ceil(Math.random() * 3)
+    var chance = Math.ceil(Math.random() * 3)
     let offset = 120
     if (chance === 1) {
         currentWeapon = "gun2"
@@ -484,6 +511,7 @@ function collectPowerUp(ship, upgrade) {
         shipUpgrade = ship3s.getFirstExists(false);
         offset = 140
     } else if (chance === 3) {
+        currentWeapon = "gun4"
         shipUpgrade = ship4s.getFirstExists(false);
     }
     shipUpgrade.anchor.setTo(.5);
